+++ sources/plugins/WES/WES.cpp
@@ -7,6 +7,7 @@
 #include <shared/plugin/yPluginApi/IDeviceRemoved.h>
 
 #include "equipments/manuallyDeviceCreationException.hpp"
+#include "equipments/tooLowRevisionException.hpp"
 #include <shared/Log.h>
 #include <boost/property_tree/xml_parser.hpp>
 
@@ -146,6 +147,11 @@ void CWES::doWork(boost::shared_ptr<yApi::IYPluginApi> api)
                {
                   request->sendError("device Name already exist");
                }
+            }
+            catch (CtooLowRevisionException& e)
+            {
+               request->sendError(e.what());
+               setPluginState(api, kserverRevisionTooLow);
             }
             catch (CManuallyDeviceCreationException& e)
             {
@@ -304,6 +310,9 @@ void CWES::setPluginState(boost::shared_ptr<yApi::IYPluginApi> api, EWESPluginSt
          api->setPluginState(yApi::historization::EPluginState::kCustom, "kAtLeastOneConnectionFaulty");
       case kmanuallyCreationDeviceFailed:
          api->setPluginState(yApi::historization::EPluginState::kCustom, "manuallyCreationError");
+         break;
+      case kserverRevisionTooLow:
+         api->setPluginState(yApi::historization::EPluginState::kCustom, "serverRevisionTooLow");
          break;
       case kRunning:
          api->setPluginState(yApi::historization::EPluginState::kRunning);