+++ sources/plugins/WES/changelog.md
@@ -1,5 +1,7 @@
 ### 1.0.1
 * change CGX files for WES revision 1 and revision 2 with last firmware
+* check the firmware revision at the startup of the plugin. The firmware revision should be greater than V0.80
+* add the third TIC counter
 * fix initialization bug with WES revision 1
 
 ### 1.0.0